# openapi.yaml
openapi: 3.0.0
info:
  title: PSL API
  contact:
    name: Email
    email: alejandro.hernandezl@infotec.mx

  description: "PSL API manual de url"
  version: 1.0.0
...
security:
  - bearerAuth: []
paths:
...
  /sentenciado:
    get:
      tags:
        - Sentenciado
      summary: Busca un sentenciado.
      description: Dado una serie de filtros, busca la información del sentenciado.
      parameters:
        - in: query
          name: sentenciadoId
          schema:
            type: string
          description: El id del sentenciado
        - in: query
          name: folio
          schema:
            type: string
          description: Folio del sentenciado
        - in: query
          name: nombre
          schema:
            type: string
          description: Nombre del sentenciado
        - in: query
          name: apellidoPaterno
          schema:
            type: string
          description: Apellido paterno del sentenciado
        - in: query
          name: apellidoMaterno
          schema:
            type: string
          description: Apellido materno del sentenciado
        - in: query
          name: paisId
          schema:
            type: number
          description: Pais materno del sentenciado
        - in: query
          name: alias
          schema:
            type: string
          description: Alias del sentenciado
        - in: query
          name: otrosNombres
          schema:
            type: string
          description: Otros nombres del sentenciado
        - in: query
          name: fechaNacimiento
          schema:
            type: string
          description: fechaNacimiento del sentenciado
        - in: query
          name: ocupacionId
          schema:
            type: number
          description: Ocupaciòn del sentenciado
        - in: query
          name: nombrePadres
          schema:
            type: string
          description: Nombre algun padre del sentenciado
        - in: query
          name: apellidoPaternoPadres
          schema:
            type: string
          description: Apellido paterno algun padre del sentenciado
        - in: query
          name: apellidoMaternoPadres
          schema:
            type: string
          description: Apellido materno padres del sentenciado
      responses:
        '200':
          description: Senteciado creado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConsultaSentenciado'
        '400':
          description: Something in the request was wrong.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Errors'
        '500':
          description: Algo en el servidor salío mal.
    post:
      tags:
        - Sentenciado
      summary: Crea a un sentencido.
      description: Crea a un sentencido. Carecteres mayores a  100 son permitidos. Folio es unico por convicto
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGeneral'
      responses:
        '200':
          description: Senteciado creado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sentenciado'
        '400':
          description: Something in the request was wrong.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Errors'
        '500':
          description: Algo en el servidor salío mal.
  /sentenciado/{id}:
    get:
      tags:
        - Sentenciado
      summary: Busca un sentenciado.
      description: Dado un ID, busca la información del sentenciado.
      parameters:
        - in: path
          name: id
          schema:
            type: integer
            example: 123
          required: true
          description: El id del sentencido.
      responses:
        '200':
          description: La información del sentenciado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sentenciado'
        '400':
          description: Algo en la petición está mal.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Errors'
        '404':
          description: El ID del sentenciado no fue encontrado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Errors'
        '500':
          description: Algo en el servidor está mal.
    put:
      tags:
        - Sentenciado
      summary: Actualiza a un sentenciado.
      description: Actualiza a un sentenciado.
      parameters:
        - in: path
          name: id
          schema:
            type: integer
            example: 123
          required: true
          description: El Id del sentenciado.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGeneral'
      responses:
        '200':
          description: Sentenciado actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sentenciado'
        '400':
          description: Algo en la petición está mal.

          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Errors'
        '404':
          description: El ID del sentenciado no fue encontrado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Errors'
        '500':
          description: Algo en el servidor está mal.
    delete:
      tags:
        - Sentenciado
      summary: Borra a un sentenciado.
      description: Borrado lógico del sentenciado.
      parameters:
        - in: path
          name: id
          schema:
            type: integer
            example: 123
          required: true
          description: El ID del sentenciado.
      responses:
        '204':
          description: El sentenciado fue borrado.
        '400':
          description: Algo en la petición está mal.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Errors'
        '404':
          description: El ID del sentenciado no fue encontrado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Errors'
        '500':
          description: Algo en el servidor está mal.
...
components:
  schemas:
    CreateGeneral:
      description: Datos nescesarios para crear un sentenciado
      required:
        - nombre
        - nacionalidad
        - estado
        - estadoCivil
        - fechaNacimiento
        - sexo
        - celular
        - email
      properties:
        nombre:
          description: Nombre del senteciado.
          type: string
          minimum: 1
          maximum: 50
          example: Juan Antonio
        apellidoMaterno:
          description: Apellido paterno.
          type: string
          minimum: 1
          maximum: 20
          example: PÃ©rez
        apellidoPaterno:
          description: Apellido materno.
          type: string
          minimum: 1
          maximum: 20
          example: GarcÃ­a
        nacionalidad:
          description: Nacionalidad del sentenciado.
          type: string
          minimum: 1
          maximum: 20
          example: MexÃ­cana
        estado:
          description: Estado provincia del sentenciado.
          type: number
          minimum: 1
          maximum: 15
          example: 15
        curp:
          description: CURP.
          type: string
          minimum: 22
          maximum: 22
          example: HELA880416HHGRZL08
        estadoCivil:
          description: Estado civil.
          type: string
          minimum: 1
          maximum: 20
          example: soltero
        alias:
          description: Alias.
          type: string
          minimum: 1
          maximum: 20
          example: el pinky
        otrosNombres:
          description: Otros nombres.
          type: string
          minimum: 1
          maximum: 20
          example: Enrique Peña
        fechaNacimiento:
          type: number
          description: Fecha de nacimiento.
          example: 1587567125458
        ocupacion:
          description: Ocupacion .
          type: string
          minimum: 1
          maximum: 20
          example: Servidor publico
        sexo:
          $ref: '#/components/schemas/Gender'
        etnia:
          description: La etnia.
          type: string
          minimum: 1
          maximum: 20
          example: Maya
        escolaridad:
          description: Escolaridad.
          type: string
          minimum: 1
          maximum: 20
          example: Maya
        telefonoFijo:
          type: number
          description: Telefono fijo del sentenciado.
          example: 1587567125458
        celular:
          type: number
          description: Numero celular.
          example: 1587567125458
        email:
          description: El email.
          type: string
          minimum: 1
          maximum: 100
          example: juan.antonio.perez.garcia@gmail.com
        ip:
          description: La ip donde se realiza la petición.
          type: string
          minimum: 22
          maximum: 22
          example: HELA880416HHGRZL08
    Gender:
      description: Sexó del sentenciado.
      type: string
      enum:
        - Femenino
        - Masculino
    Sentenciado:
      description: Datos de la pantalla generales.
      required:
        - nombre
        - nacionalidad
        - estado
        - estadoCivil
        - fechaNacimiento
        - sexo
        - celular
        - email
      properties:
        id:
          description: Id del sentenciado
          type: integer
          example: 123
        folio:
          description: Folio del sentenciado
          type: string
          minimum: 1
          maximum: 50
          example: MEJF84020515HMEX017
        nombre:
          description: Nombre del senteciado.
          type: string
          minimum: 1
          maximum: 50
          example: Juan Antonio
        apellidoMaterno:
          description: Apellido paterno.
          type: string
          minimum: 1
          maximum: 20
          example: PÃ©rez
        apellidoPaterno:
          description: Apellido materno.
          type: string
          minimum: 1
          maximum: 20
          example: GarcÃ­a
        nacionalidad:
          description: Nacionalidad del sentenciado.
          type: string
          minimum: 1
          maximum: 20
          example: MexÃ­cana
        estado:
          description: Estado provincia del sentenciado.
          type: string
          minimum: 1
          maximum: 20
          example: Hidalgo
        curp:
          description: CURP.
          type: string
          minimum: 22
          maximum: 22
          example: HELA880416HHGRZL08
        estadoCivil:
          description: Estado civil.
          type: string
          minimum: 1
          maximum: 20
          example: soltero
        alias:
          description: Alias.
          type: string
          minimum: 1
          maximum: 20
          example: el pinky
        otrosNombres:
          description: Otros nombres.
          type: string
          minimum: 1
          maximum: 20
          example: Enrique Peña
        fechaNacimiento:
          type: number
          description: Fecha de nacimiento.
          example: 1587567125458
        ocupacion:
          description: Ocupacion .
          type: string
          minimum: 1
          maximum: 20
          example: Servidor publico
        sexo:
          $ref: '#/components/schemas/Gender'
        etnia:
          description: La etnia.
          type: string
          minimum: 1
          maximum: 20
          example: Maya
        escolaridad:
          description: Escolaridad.
          type: string
          minimum: 1
          maximum: 20
          example: Maya
        telefonoFijo:
          type: number
          description: Telefono fijo del sentenciado.
          example: 1587567125458
        celular:
          type: number
          description: Numero celular.
          example: 1587567125458
        email:
          description: El email.
          type: string
          minimum: 1
          maximum: 100
          example: juan.antonio.perez.garcia@gmail.com
        dateCreated:
          type: number
          description: The date when the user was created.
          example: 1587567125458
    ConsultaSentenciado:
      description: Datos de la pantalla consulta.
      properties:
        id:
          description: Id del sentenciado
          type: integer
          example: 123
        folio:
          description: Folio del sentenciado
          type: string
          minimum: 1
          maximum: 50
          example: MEJF84020515HMEX017
        nombre:
          description: Nombre del senteciado.
          type: string
          minimum: 1
          maximum: 50
          example: Juan Antonio
        apellidoMaterno:
          description: Apellido paterno.
          type: string
          minimum: 1
          maximum: 20
          example: PÃ©rez
        apellidoPaterno:
          description: Apellido materno.
          type: string
          minimum: 1
          maximum: 20
          example: GarcÃ­a
        procesos:
          description: lista de porcesos.
          type: string
          minimum: 1
          maximum: 100
          example: proceso, proceso
...
    Errors:
      description: A list of errors
      type: array
      items:
        $ref: '#/components/schemas/Error'
    Error:
      description: A detailed description of an error
      properties:
        title:
          type: string
          description: A brief reason of the error.
          example: "Error title"
        code:
          type: string
          description: A unique code of the error.
          example: "error.unique.code"
        detail:
          type: string
          description: A detailed description of the error.
          example: "Something in your request was wrong or was not found"
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: You can get the access token through the "/login" endpoint.